# docs/openapi/openapi.v0.12.0.yaml
openapi: 3.0.3
info:
  title: SaaS Boilerplate API
  version: 0.12.0
  description: Organisation-first SaaS boilerplate API. v0.12.0 completes the OIDC
    / SSO build by activating authorization-code + PKCE start/callback/logout routes,
    OIDC discovery, token exchange, RS256 id_token validation, auto-provisioning,
    group-to-role mapping and audit/security events for SSO login.
servers:
- url: http://localhost:4000/api
  description: Local development API
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Generic bearer auth scheme.
    UserSessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT user_session_token
      description: Short-lived logged-in user token used by the frontend and Postman
        during development.
    PersonalAccessTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT personal_access_token
      description: Long-lived user-owned scoped token for Postman, scripts and user
        automation.
    OrganisationApiTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT organisation_api_token
      description: Organisation-owned token for external/system-to-system integrations.
        Uses external audience and external:* scopes.
    ServiceAccountTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: Service Account Token
      description: Reserved for later service identity requirements.
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            details:
              type: object
...
paths:
...
  /org/sso-config/test:
    post:
      tags:
      - SSO
      summary: Validate organisation SSO config
      description: 'Runs local validation against saved SSO provider configuration.
        Authentication: UserSessionAuth or PersonalAccessTokenAuth. Required permission/scope:
        organisation.sso.view.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Validate organisation SSO config succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '201':
          description: Validate organisation SSO config created resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /org/personal-access-tokens:
    get:
      tags:
      - Internal
      summary: List personal access tokens
      description: 'Lists current user''s personal access tokens. Authentication:
        UserSessionAuth or PersonalAccessTokenAuth. Required permission/scope: tokens.personal.manage.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      responses:
        '200':
          description: List personal access tokens succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Internal
      summary: Create personal access token
      description: 'Creates a user-owned long-lived scoped token. Raw token is returned
        once. Authentication: UserSessionAuth or PersonalAccessTokenAuth. Required
        permission/scope: tokens.personal.manage.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Create personal access token succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '201':
          description: Create personal access token created resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /org/personal-access-tokens/{tokenId}:
    delete:
      tags:
      - Internal
      summary: Revoke personal access token
      description: 'Revokes a user-owned personal access token. Authentication: UserSessionAuth
        or PersonalAccessTokenAuth. Required permission/scope: tokens.personal.manage.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Revoke personal access token succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '201':
          description: Revoke personal access token created resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...